requirements.txt
Pillow
mypy
types-Pillow
